requirements.txt
python-dotenv
load_dotenv
requests
PdfReader
nltk
chromadb
PyPDF2